Add files via upload
The original dialog window was substituted with a "point and click" selection selection of Directory, followed by selection of function to be run (Convert Masks to ROIs, Check ROIs, Measure). Upon selection a function, the macro runs it on all directories within the selected directory recursively. The macro can then be run again with the choice of the next function. Since only the "Check ROIs" function requires human intervention, this is a more time-efficient way to process the data.